+++ src/main/java/callprotector/spring/domain/callchat/controller/CallChatLogController.java
@@ -1,6 +1,6 @@
 package callprotector.spring.domain.callchat.controller;
 
-import callprotector.spring.domain.chat.dto.response.CallChatLogResponseDTO;
+import callprotector.spring.domain.callchat.dto.response.CallChatLogResponseDTO;
 import callprotector.spring.domain.chat.entity.CallChatSession;
 import callprotector.spring.domain.chat.service.CallChatLogService;
 import callprotector.spring.domain.chat.service.CallChatSessionService;
