+++ src/main/java/callprotector/spring/domain/chat/service/CallChatLogService.java
@@ -1,7 +1,6 @@
 package callprotector.spring.domain.chat.service;
 
-import callprotector.spring.domain.chat.dto.response.CallChatLogResponseDTO;
-import callprotector.spring.domain.chat.entity.CallChatLog;
+import callprotector.spring.domain.callchat.dto.response.CallChatLogResponseDTO;
 
 import java.util.List;
